+++ compiler/Options.hs
@@ -42,6 +42,7 @@ data Options
         , comm_enabled :: Bool
         , grpc_enabled :: Bool
         , alloc_ctors_enabled :: Bool
+        , type_aliases_enabled :: Bool
         , service_inheritance_enabled :: Bool
         }
     | Cs
@@ -97,6 +98,7 @@ cpp = Cpp
     , comm_enabled = False &= explicit &= name "comm" &= help "Generate comm definitions"
     , grpc_enabled = False &= explicit &= name "grpc" &= help "Generate gRPC definitions"
     , alloc_ctors_enabled = False &= explicit &= name "alloc-ctors" &= help "Generate constructors with allocator argument"
+    , type_aliases_enabled = False &= explicit &= name "type-aliases" &= help "Generate type aliases"
     , service_inheritance_enabled = False &= explicit &= name "enable-service-inheritance" &= help "Enable service inheritance syntax in IDL"
     } &=
     name "c++" &=